requirements.txt
matplotlib==3.7.2
loompy==3.0.7
pandas==2.0.3
pymde==0.1.18
seaborn==0.12.2
scanpy==1.9.3
scib-metrics==0.3.3
scikit-learn==1.3.0
scikit-misc==0.2.0
scvi-tools==0.20.3
tensorflow-datasets==4.9.2
tensorflow-probability==0.20.0